Updates glium dependency to 0.30
This change updates to the latest version of glium (0.30), which brings
glutin to version 0.27 and winit to 0.25. This fixes known issue
[#1773](https://github.com/rust-windowing/winit/issues/1773) in winit
0.19 on Linux X11.

The key stubstanitive change to the examples is the change from glutin's
`EventsLoop` to `EventLoop`. Instead of a user-defined while loop where
`EventsLoop::poll_events` is called, `EventLoop` has switched to a run
function which is event-driven by glutin's core. This slightly changes
the program structure.

The other changes are minor, as glutin moved a few data structures into
modules instead of having everything visible at the top-level.